requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
httpx>=0.27.2
pydantic>=2.7.1
pyjwt[crypto]>=2.8.0
sqlalchemy>=2.0.31
python-dotenv>=1.0.1
bcrypt>=4.2.0
tiktoken>=0.7.0
pyyaml>=6.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
Pillow>=10.0.0
openpyxl>=3.1.0
